hackathon.02/phase-1-console/requirements.txt
pytest>=7.4
python-dateutil>=2.8
